CCSTPES/teacher_eval_project/requirements.txt
Django==4.2.0
psycopg2-binary==2.9.6
python-decouple==3.8
Pillow==10.0.0
django-storages==1.14.2
boto3==1.28.0
gunicorn==21.2.0
whitenoise==6.5.0
django-cors-headers==4.2.0
